Fixed creation of default user

+++ user/migrations/0002_add_user.py
@@ -1,3 +1,4 @@
+from django.contrib.auth.hashers import make_password
 from django.db import migrations
 
 from user.models import User
@@ -6,7 +7,7 @@ from user.models import User
 def add_user(apps, schema_editor):
     user = User(
         email="admin@example.com",
-        password="test_admin12345",
+        password=make_password("test_admin12345"),
         is_staff=True,
         is_superuser=True,
     )
@@ -14,11 +15,11 @@ def add_user(apps, schema_editor):
 
 
 def remove_user(apps, schema_editor):
-    User.objects.get(email="admin@example.com").delete()
+    if User.objects.filter(email="admin@example.com").exists():
+        User.objects.get(email="admin@example.com").delete()
 
 
 class Migration(migrations.Migration):
-
     dependencies = [
         ("user", "0001_initial"),
     ]
